+++ crates/rs7-wasm/src/utils.rs
@@ -1,0 +1,30 @@
+//! Utility functions for WebAssembly
+
+use wasm_bindgen::prelude::*;
+
+/// Set up better panic messages in wasm
+pub fn set_panic_hook() {
+    #[cfg(feature = "console_error_panic_hook")]
+    console_error_panic_hook::set_once();
+}
+
+/// Log to browser console
+#[wasm_bindgen]
+extern "C" {
+    #[wasm_bindgen(js_namespace = console)]
+    pub fn log(s: &str);
+
+    #[wasm_bindgen(js_namespace = console)]
+    pub fn warn(s: &str);
+
+    #[wasm_bindgen(js_namespace = console)]
+    pub fn error(s: &str);
+}
+
+/// Macro for logging to console
+#[macro_export]
+macro_rules! console_log {
+    ($($t:tt)*) => {
+        $crate::utils::log(&format!($($t)*))
+    };
+}